step 4 - 유효성 검사 클래스 분리
