+++ src/main/java/racingcar/racing/Racing.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static java.lang.System.out;
+import static racingcar.validate.CarValidate.validateCarsName;
 import static racingcar.view.OutputView.printCarStatus;
 import static racingcar.view.OutputView.resultView;
 
@@ -20,20 +21,6 @@ public class Racing {
             out.println();
         }
         resultView(cars);
-    }
-
-    public static String[] validateCarsName(String carsName) {
-        String[] carsArray = carsName.split(",");
-        for(String carName : carsArray) {
-            validateCarName(carName);
-        }
-        return carsArray;
-    }
-
-    public static void validateCarName(String carName) {
-        if(carName.length() > 5 || carName.length() < 1) {
-            throw new IllegalArgumentException("자동차 이름은 최소 1자 이상 5자 이하여야 합니다");
-        }
     }
 
     public static List<Car> initCars(String[] carsName) {
