+++ src/main/java/racingcar/validate/CarValidate.java
@@ -1,0 +1,17 @@
+package racingcar.validate;
+
+public class CarValidate {
+    public static String[] validateCarsName(String carsName) {
+        String[] carsArray = carsName.split(",");
+        for(String carName : carsArray) {
+            validateCarName(carName);
+        }
+        return carsArray;
+    }
+
+    public static void validateCarName(String carName) {
+        if(carName.length() > 5 || carName.length() < 1) {
+            throw new IllegalArgumentException("자동차 이름은 최소 1자 이상 5자 이하여야 합니다");
+        }
+    }
+}
