+++ src/test/java/racingcar/racing/RacingTest.java
@@ -2,14 +2,11 @@ package racingcar.racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.car.Car;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static racingcar.racing.Racing.*;
 
@@ -18,16 +15,9 @@ public class RacingTest {
 
     @BeforeEach
     void init() {
-
         cars.add(new Car("애플"));
         cars.add(new Car("삼성"));
         cars.add(new Car("구글"));
-    }
-
-    @Test
-    void 경주시작_임의값_성공() {
-        String carsName = "애플,삼성,구글";
-        racingStart(carsName, 5);
     }
 
     @Test
@@ -37,21 +27,13 @@ public class RacingTest {
     }
 
     @Test
-    void 자동차_전진_성공() {
-        moveCarForward(cars);
+    void 경주시작_임의값_성공() {
+        String carsName = "애플,삼성,구글";
+        racingStart(carsName, 5);
     }
 
     @Test
-    void 자동차이름들_유효성_검사() {
-        String carsName = "애플,삼성,구글";
-        String[] carsArray = validateCarsName(carsName);
-        assertEquals(3, carsArray.length);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"", "123456"})
-    void 자동차이름_한글자이상_다섯글자이하_검사(String name) {
-        assertThatIllegalArgumentException().isThrownBy(() -> validateCarName(name))
-                .withMessage("자동차 이름은 최소 1자 이상 5자 이하여야 합니다");
+    void 자동차_전진_성공() {
+        moveCarForward(cars);
     }
 }
